restore root scope with predeclared `use`
Motive: lower learning curve and nicer aesthetics of mixed imports
at the top level of every module. It's much simpler to say "the only
predeclared name is `use`" than to explain contextual live values.
Mixing `.use` with `import` and `declare` has questionable aesthetics.
In addition, contextual live values / contextual sub-macros are an
experimental feature which is still under consideration, and may be
removed in the future, while having a root scope with one predeclared
identifier is less questionable and much less likely to be removed or
changed.