simpler macro internals; in-place recursive macroing; quoting fixes
This commit drops support for "source nodes" which were used to trace
replacement nodes to original nodes. Instead, we simply copy spans.
This means we lose support for source node tracing, at least for now.
The motive is that source node references tended to cause complexity
in various parts of the system in ways that were hard to manage. Even
though we prevented cycles, the existence of source node relations
sometimes led to infinite recursion anyway. We may reintroduce a similar
feature in the future, but for now, avoiding it seems like a better
tradeoff.

This commit changes recursive macroing from out-of-place to in-place.
This means when a node returns a node that returns a node and so on,
each new node is attached to its parent _before_ we invoke its own
macro method. This allows replacement nodes to correctly use methods
such as `.isStatement` which require the child-parent relation to be
bilateral rather than unilateral. We also now automatically validate
bilateral relations where relevant.

This commit changes how `ListMacro` handles children obtained from a
generic node list. Previously, it copied all children as-is. Now, it
drops the first child, assuming that it represents an identifier that
resolved to the current macro. This is done to simplify programmatic
construction of macro nodes, which is now happening in more and more
places. One example is `ModuleNodeList..reqAutoImport`, which
constructs an import statement programmatically. In addition, with
this change, function-style macros and `ListMacro` subclasses handle
their parameters / children symmetrically: index 0 of the original
node list is argument 0 for function-style macros, and child 0 for
class-style macros.

This commit adds missing support for "own" names and values for some
nodes that were previously lacking it, such as `Num`, `Str`, `Ident`.
Previously, they outsourced this data to spans. With the removal of
source nodes, we simply reuse spans from source nodes, giving them to
replacement nodes. This means in the general case, a node's span may
represent arbitrary code, and can't be used to derive the node's data
or used to compile the node. Everything required for correctly
constructing and compiling a node must be stored in own fields other
than the node's span. This commit also ensures that we correctly
compile those fields when quoting. Decoupling nodes from spans also
makes it easier to construct nodes programmatically.