improvements to macroing and compilation
Consolidated the macroing of symbols with the macroing of all other nodes.
Symbols that refer to runtime-only values behave as before. Symbols that
refer to compile-time values are now exactly equivalent to the values that
they reference. Compile-time symbolic references are no longer limited to
functions. They may refer to arbitrary nodes.

Consolidated the macroing of functions and function-referencing symbols in
non-call positions with the macroing of arbitrary objects. In both cases,
the referenced value may implement the `.macro` method to support macroing
in arbitrary positions. The `.default` method that was specially supported
by symbols is no longer used.

Most "operator" macros can now also be used in arbitrary expression positions,
compiling to tiny inline functions in such cases. When used in call positions,
they compile to pure operator expressions as before.

Added support for regexp compilation.

Various additions to the prelude and debug modules. Various fixes and tweaks.