revise class declarations; lexical fields and methods
Class fields are now declared with `let`. Class methods are now declared with
`func`. Fields with computed names are declared with `set`. All of these forms
now support contextual overrides, which are used by classes.

Fields declared with `let` and methods declared with `func` are now declared
lexically, and may be used simply by name, without the preceding `this.`. This
behavior is similar to some compiled OO languages such as Java.

In the immediate body of a class definition, fields and methods are static.
Declaring instance fields and methods requires an additional opt-in via the
`prototype` form. This ordering is different from JS and Java, where instance
fields and methods are the default, and static requires an explicit opt-in.
This choice massively simplifies the implementation of lexical scoping for
fields and methods, and avoids some field reference ambiguities that would be
otherwise present. It also better aligns with how fields and methods are
accessed on already-defined classes.

Removed the `static` macro namespace. Static blocks in class definitions are
now written with the `do` macro, via a contextual override.

The macro `dict` now supports object spread, via `...` in key position.

`get` can be used as `.`, and `getOpt` can be used as `.?`, if the prelude
module is imported as `*`.

Other misc corrections and tweaks.

Minor compilation performance regressions.