Scheme-style declarations, explicit list deconstructions
This commit changes `func` declarations to the Scheme style, where the
function name and the parameters are placed in the same list. This also
changes list deconstructions in function parameters and variable
declarations to the explicit style, where each deconstruction must
begin with a reference to the list macro.

This allows symmetry / consistency between deconstructions in
declarations and assignments, which were previously inconsistent
with each other, and minimizes situations where the first element
in a list is a declared name rather than a callable.

For consistency, this also changes `class` to the same style,
and removes the `extend` macro.

Various other minor additions and tweaks.