support getter functions / property getters in classes
Support for property setters in classes is delayed because we currently
don't allow redundant declarations of the same name, and it seems like
getter and setter pairs would require that. Support for property setters
in object literals is delayed because our current "dict" macro does not
support special features such as methods, getters, setters, or shorthand
properties. This may be implemented later.