alternative object literal macro with `set` `let` `func`
This introduces another object literal macro `obj`, in addition to the
object literal macro `dict`. The previously-existing `dict` is better
suited for the delimiter based dialect because key-value entries are
represented inline, without additional delimiters. The new `obj` is
better suited for the indent based dialect, where using `dict` requires
the use of the splicer character `'`. Declaring fields with `let` and
`func` is visually cleaner and consistent with regular variables and
functions outside of objects.