mixin imports use `...`, not `*`
The use of `*` was inertia from JS, where it's also used in imports.
However, the conceptual role of `*` in JS imports is similar to the
rest operation in dicts, while its conceptual role in Jisp `use` is
similar to the spread operation in dicts. Using the spread operator
seems like a better fit. Also fewer "special" symbols to learn.