more consistent rest/spread in lists
Previously, rest and spread in lists, including parameter lists, were
asymmetric. Rest was implemented by treating a specific symbol as an
unary prefix keyword, while spread was a regular macro. This required
special-case logic in deconstructions, and caused declaration and usage
to be inconsistent:

    [func [blah ... rest]]
    [blah [... args]]

This commit replaces special-case rest-parameter-related logic in list
deconstructions with an override for the `spread` macro, similar to the
already-present override for the `list` macro, and expects user code to
call that override in deconstructions:

    [func [blah [... rest]]]
    [blah [... args]]

This has the obvious downside of more characters and more visual noise.
But it buys us simplicity, both internal and conceptual, and allows for
symmetric declaration and usage.

Spread in the `dict` macro still relies on special treatment of `...`
in the key position. We could generalize this by detecting any symbol
that references the `spread` macro, or by treating macros referenced
in key positions as macro calls with the corresponding value as the
input. But it's hard to justify the added code and the overhead when
this would have no effect on any real user code. This commit changes
only rest support, not spread support.